+++ NEWS.md
@@ -8,14 +8,23 @@ editor_options:
 
 # ggpp 0.5.7
 
--   Fix bug in `geom_point_s()`, `alpha_target = "point"` ignored.
--   Update `geom_point_s()` adding parameter `move.point` to allow its use to highlight point at the original position with an arrow with its start given
-by the action of a position function.
--   Change in `geom_point_s()` the end at which the tip of the arrow is drawn to its most frequently used direction, even if opposite to the default for
+-   This version depends on 'ggplot2' >= 3.5.0.
+-   Update `geom_point_s()` adding parameter `move.point` to allow its use to 
+highlight points at the original position with an arrow with its start given by 
+the action of a position function.
+-   Change in `geom_point_s()` the end at which the tip of the arrow is drawn
+to its most frequently used direction, even if opposite to the default for
 `geom_text_s()` and `geom_label_s()`.
 -   Update `geom_text_s()`, `geom_label_s()`. `geom_text_pairwise()` and 
 `geom_label_pairwise()` adding parameter `size.unit` tracking change in 
 `geom_text()` and `geom_label()` in 'ggplot2' 3.5.0.
+-   Update `geom_text_s()`, `geom_label_s()`. `geom_text_pairwise()` and 
+`geom_label_pairwise()` so that graphic elements in the keys match those in
+the plot, even when using the additional features not availablr in 'ggplot2'.
+-   Fix in `geom_text_s()`, `geom_label_s()`. `geom_text_pairwise()` and 
+`geom_label_pairwise()` an infrequent problem with incomplete guides in 
+'ggplot2' 3.5.0.
+-   Fix bug in `geom_point_s()`, `alpha_target = "point"` ignored.
 
 # ggpp 0.5.6